Consolidate duplicate DataFrame and GroupedData implementations
- Remove duplicate MockDataFrame implementations (core/dataframe.py)
- Remove duplicate MockGroupedData implementations (grouped_data.py)
- Consolidate into single main DataFrame with all features
- Add missing toDuckDB() method and schema properties to main DataFrame
- Update all imports to use consolidated implementations
- Fix DuckDB type mapping (FloatType -> DOUBLE, LongType -> BIGINT)
- Remove slow test_join_performance test
- All 407 tests passing with improved performance

+++ tests/unit/test_dataframe_duckdb_integration.py
@@ -7,7 +7,7 @@ Tests the new toPandas() optional dependency and toDuckDB() analytical features.
 import pytest
 import duckdb
 from unittest.mock import patch, MagicMock
-from mock_spark.dataframe.core.dataframe import MockDataFrame
+from mock_spark.dataframe.dataframe import MockDataFrame
 from mock_spark.spark_types import MockStructType, MockStructField, StringType, IntegerType, DoubleType
 
 
@@ -233,7 +233,7 @@ class TestDataFrameDuckDBIntegration:
         # Test type mapping method directly
         assert df._get_duckdb_type(StringType()) == "VARCHAR"
         assert df._get_duckdb_type(IntegerType()) == "INTEGER"
-        assert df._get_duckdb_type(LongType()) == "INTEGER"
+        assert df._get_duckdb_type(LongType()) == "BIGINT"
         assert df._get_duckdb_type(DoubleType()) == "DOUBLE"
         assert df._get_duckdb_type(FloatType()) == "DOUBLE"
         assert df._get_duckdb_type(BooleanType()) == "BOOLEAN"
